lib: rename enrichers to plugins and create structure
